pyproject.toml
[tool.poetry]
name = "centralized-kyc"
version = "0.1.0"
description = ""
authors = ["shalin-devkota <shalin.devkota01@gmail.com>"]
readme = "README.md"
packages = [{include = "centralized_kyc"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
